Unseal hierarchy
